# swagger.yaml
openapi: 3.0.0
info:
  title: API Documentation task for PsiBorg Technologies
  description: API Satya Singh
  version: 1.0.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PushNotificationRequest:
      type: object
      properties:
        deviceToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2Nzc2NzEyYjJlZjQxYTdmYTczNzBmNzUiLCJ1c2VybmFtZSI6IlNhdHlhU2luZ2giLCJlbWFpbCI6InNhdHlhcHJha2FzaHNpbmdoa2FzaWFAZ21haWwuY29tIiwiaWF0IjoxNzM2MDc0NzEwLCJleHAiOjE3MzY0MzQ3MTB9.PEkemM_4nZKQWgQO0GYcf-QDpSybj_54MObY2NS9Ecc"
        title:
          type: string
          example: "Hello, User!"
        body:
          type: string
          example: "This is your first push notification."
    PushNotificationResponse:
      type: object
      properties:
        message:
          type: string
          example: "Push notification sent successfully"
paths:
...
  /api/admin/pushNotification:
    post:
      tags:
        - Admin
      summary: Send a push notification
      description: Sends a push notification to a user's device using a provided device token.
      operationId: sendPushNotification
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PushNotificationRequest'
            example:
              deviceToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2Nzc2NzEyYjJlZjQxYTdmYTczNzBmNzUiLCJ1c2VybmFtZSI6IlNhdHlhU2luZ2giLCJlbWFpbCI6InNhdHlhcHJha2FzaHNpbmdoa2FzaWFAZ21haWwuY29tIiwiaWF0IjoxNzM2MDc0NzEwLCJleHAiOjE3MzY0MzQ3MTB9.PEkemM_4nZKQWgQO0GYcf-QDpSybj_54MObY2NS9Ecc"
              title: "Hello, User!"
              body: "This is your first push notification."
      responses:
        '200':
          description: Push notification sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PushNotificationResponse'
        '400':
          description: Bad Request (Invalid or missing input)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid device token or notification content."
        '401':
          description: Unauthorized (Invalid or missing Bearer token)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Unauthorized. Please provide a valid Bearer token."
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Something went wrong, please try again later."
...